api changes

+++ controllers/library-controllers.js
@@ -3,7 +3,7 @@ const Role = require("../Models/Role");
 const File = require("../Models/File");
 const Settings = require("../Models/Settings");
 const Library = require("../Models/Library");
-
+const librarySchema = require("../Models/Library");
 const getArticles = async (req, res) => {
   try {
     let foundArticles = await Library.find({});
@@ -45,26 +45,28 @@ const getSingleArticle = async (req, res) => {
   const { articleId } = req.body;
 
   try {
-    let foundArticle = await Library.findOne({ _id: articleId });
-    if (foundArticle) {
-      res.json({
-        serverError: 0,
-        message: "Article found",
-        data: {
-          success: 1,
-          article: foundArticle,
-        },
-      });
-      return;
-    } else {
-      res.json({
-        serverError: 0,
-        message: "No article found",
-        data: {
-          success: 0,
-        },
-      });
-      return;
+    if (articleId) {
+      let foundArticle = await Library.findById({ _id: articleId });
+      if (foundArticle) {
+        res.json({
+          serverError: 0,
+          message: "Article found",
+          data: {
+            success: 1,
+            article: foundArticle,
+          },
+        });
+        return;
+      } else {
+        res.json({
+          serverError: 0,
+          message: "No article found",
+          data: {
+            success: 0,
+          },
+        });
+        return;
+      }
     }
   } catch (err) {
     console.log(err);
@@ -80,18 +82,31 @@ const getSingleArticle = async (req, res) => {
 };
 
 const addArticle = async (req, res) => {
+  let imageFiles = [];
+  if (req?.files?.length > 0) {
+    console.log(req.files, "here are the files");
+    imageFiles = req.files.map((file) => file.path);
+  }
   console.log(req.body);
-  const { name, description, image, author, authorName, authorImage } =
-    req.body;
+  const {
+    title,
+    description,
+    image,
+
+    authorName,
+    //authorImage
+    date,
+  } = req.body;
   try {
     let newArticle = new Library({
-      name,
+      title,
       description,
-      image,
+      image: imageFiles,
       author: {
         authorName,
-        authorImage,
+        //authorImage,
       },
+      date,
     });
     newArticle.save((err, data) => {
       if (err) {
@@ -115,23 +130,80 @@ const addArticle = async (req, res) => {
   }
 };
 
+// const updateArticle = async (req, res) => {
+//   console.log(req.body);
+//   let imageFiles = [];
+//   if (req?.files?.length > 0) {
+//     console.log(req.files, "here are the files");
+//     imageFiles = req.files.map((file) => file.path);
+//   }
+//   const {
+//     articleId,
+//     title,
+//     description,
+//     image,
+//     author,
+//     authorName,
+//     //authorImage,
+//     date,
+//   } = req.body;
+
+//   try {
+//     Library.updateOne(
+//       { _id: articleId },
+//       {
+//         $set: { ...req.body, image: imageFiles },
+//       },
+//       (error, data) => {
+//         if (error) {
+//           return console.log(error);
+//         } else {
+//           console.log(data, "data");
+//           if (!data) {
+//             res.json({
+//               serverError: 0,
+//               message: "Not updated",
+//               success: 0,
+//             });
+//           } else {
+//             res.json({
+//               serverError: 0,
+//               message: "Article updated",
+//               success: 1,
+//             });
+//           }
+//         }
+//       }
+//     );
+//   } catch (err) {
+//     console.log(err);
+//   }
+// };
+
 const updateArticle = async (req, res) => {
   console.log(req.body);
+  let imageFiles = [];
+  if (req?.files?.length > 0) {
+    console.log(req.files, "here are the files");
+    imageFiles = req.files.map((file) => file.path);
+  }
+
   const {
-    articleId,
-    name,
+    _id,
+    title,
     description,
     image,
-    author,
     authorName,
-    authorImage,
+    //authorImage,
+    date,
   } = req.body;
+  let updateImage = imageFiles.length > 0 ? imageFiles : image;
 
   try {
-    Library.updateOne(
-      { _id: articleId },
+    librarySchema.findByIdAndUpdate(
+      _id,
       {
-        $set: { ...req.body },
+        $set: { ...req.body, image: updateImage },
       },
       (error, data) => {
         if (error) {
@@ -162,28 +234,31 @@ const updateArticle = async (req, res) => {
 const deleteArticle = async (req, res) => {
   console.log(req.body);
   const { articleId } = req.body;
+
   try {
-    let articleFound = await Library.findByIdAndRemove({ _id: articleId });
-    console.log(articleFound);
-    if (articleFound) {
-      res.json({
-        serverError: 0,
-        message: "Article deleted successfully.",
-        data: {
-          success: 1,
-          //article: articleFound,
-        },
-      });
-      return;
-    } else {
-      res.json({
-        serverError: 0,
-        message: "Error in deleting Article",
-        data: {
-          success: 0,
-        },
-      });
-      return;
+    if (articleId) {
+      let articleFound = await Library.findByIdAndRemove({ _id: articleId });
+      console.log(articleFound);
+      if (articleFound) {
+        res.json({
+          serverError: 0,
+          message: "Article deleted successfully.",
+          data: {
+            success: 1,
+            //article: articleFound,
+          },
+        });
+        return;
+      } else {
+        res.json({
+          serverError: 0,
+          message: "Error in deleting Article",
+          data: {
+            success: 0,
+          },
+        });
+        return;
+      }
     }
   } catch (err) {
     console.log(err);
